+++ src/modules/series/SeriesTypes.ts
@@ -15,6 +15,7 @@ export interface PointT {
     rank: number;
     perf: number | null;
     testTaskKey: string;
+    parentTaskKeys: string[];
 }
 
 export interface SerieT extends SerieFeaturesT {