+++ src/modules/series/SeriesUtils.ts
@@ -79,6 +79,7 @@ export function buildSeries(
                     ? testtuple.test.perf
                     : null,
             testTaskKey: testtuple.key,
+            parentTaskKeys: testtuple.parent_task_keys,
         };
 
         const serieFeatures = buildSerieFeatures(
@@ -186,6 +187,7 @@ export function buildAverageSerie(series: SerieT[]): SerieT | null {
             rank: parseInt(rank),
             perf: average(perfs),
             testTaskKey: `average for rank ${rank}`,
+            parentTaskKeys: [],
         }));
 
     return {